saving

+++ FantasyGolfball/Models/Scoring.cs
@@ -8,5 +8,5 @@ public class Scoring
     public int PlayerId { get; set; }
     public int SeasonYear { get; set; }
     public int SeasonWeek { get; set; }
-    public double Points { get; set; }
+    public float Points { get; set; }
 }
